:truck: chore(api): move cpt certification export service to `src`
