+++ api/tests/unit/domain/services/cpf/cpf-certification-xml-export-service_test.js
@@ -4,7 +4,7 @@ import dayjs from 'dayjs';
 import timezone from 'dayjs/plugin/timezone.js';
 import utc from 'dayjs/plugin/utc.js';
 
-import * as cpfCertificationXmlExportService from '../../../../../lib/domain/services/cpf-certification-xml-export-service.js';
+import * as cpfCertificationXmlExportService from '../../../../../src/certification/session-management/domain/services/cpf-certification-xml-export-service.js';
 import { domainBuilder, expect, sinon, streamToPromise } from '../../../../test-helper.js';
 
 const { PassThrough } = stream;
